requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
boto3==1.34.82
python-dotenv==1.0.1

opencv-contrib-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.3.0

mangum==0.17.0

# optional for future
scikit-learn==1.4.2